chore(popup): Moved popup-stack out of labs-react * Simplified API - removed ephemeral vs persistent popup types * Fixed tests and dependency issues
