+++ modules/_labs/menu/react/package.json
@@ -46,7 +46,8 @@
   ],
   "devDependencies": {
     "@material-ui/core": "^4.9.7",
-    "@workday/canvas-system-icons-web": "^1.0.20"
+    "@workday/canvas-system-icons-web": "^1.0.20",
+    "@workday/canvas-kit-react-popup": "^4.0.0-beta.3"
   },
   "dependencies": {
     "@emotion/core": "^10.0.28",
