--- modules/_labs/popup-stack/react/README.md
@@ -1,36 +1,0 @@
-# Canvas Kit React Popup Stack
-
-<a href="https://github.com/Workday/canvas-kit/tree/master/modules/_labs/README.md">
-  <img src="https://img.shields.io/badge/LABS-alpha-orange" alt="LABS: Alpha" />
-</a>  This component is work in progress and currently in pre-release.
-
-Stack for managing popup UIs to coordinate global concerns like escape key handling and rendering order
-
-## Installation
-
-```sh
-yarn add @workday/canvas-kit-labs-react-popup-stack
-```
-
-## Usage
-
-```tsx
-import * as React from 'react';
-import PopupStack from '@workday/canvas-kit-labs-react-popup-stack';
-
-<PopupStack />;
-```
-
-## Static Properties
-
-> None
-
-## Component Props
-
-### Required
-
-> None
-
-### Optional
-
-> None
