+++ modules/common/react/package.json
@@ -55,7 +55,6 @@
   "dependencies": {
     "@emotion/core": "^10.0.28",
     "@emotion/styled": "^10.0.27",
-    "@popperjs/core": "^2.1.1",
     "@workday/canvas-colors-web": "^0.17.13",
     "@workday/canvas-kit-react-core": "^4.0.0-beta.3",
     "chroma-js": "^2.1.0",
