+++ modules/modal/react/package.json
@@ -47,8 +47,7 @@
   "dependencies": {
     "@emotion/core": "^10.0.28",
     "@emotion/styled": "^10.0.27",
-    "@workday/canvas-kit-react-popup": "^4.0.0-beta.3",
-    "focus-trap-js": "1.0.8"
+    "@workday/canvas-kit-react-popup": "^4.0.0-beta.3"
   },
   "peerDependencies": {
     "react": ">= 16.8 < 17"
