+++ modules/popup-stack/stories/stories.tsx
@@ -1,8 +1,8 @@
-/// <reference path="../../../../../typings.d.ts" />
-import React, {useEffect} from 'react';
+/// <reference path="../../../typings.d.ts" />
+import React from 'react';
 import withReadme from 'storybook-readme/with-readme';
 
-import {PopupStack} from '@workday/canvas-kit-labs-react-popup-stack';
+import {PopupStack} from '@workday/canvas-kit-popup-stack';
 import README from '../README.md';
 
 export default {
