+++ modules/popup-stack/tsconfig.json
@@ -1,5 +1,5 @@
 
 {
-  "extends": "../../../../tsconfig.json",
+  "extends": "../../tsconfig.json",
   "exclude": ["node_modules", "ts-tmp", "dist", "spec", "stories"]
 }
