+++ modules/popup/react/lib/useOutsideClick.ts
@@ -1,16 +1,13 @@
 import React from 'react';
 
-import {PopupStack} from '@workday/canvas-kit-labs-react-popup-stack';
+import {PopupStack} from '@workday/canvas-kit-popup-stack';
 
 export const useOutsideClick = <E extends HTMLElement>(
   ref: React.RefObject<E>,
   onClose: () => void
 ) => {
   const onClick = (event: MouseEvent) => {
-    if (
-      PopupStack.isTopmost(ref.current!, 'ephemeral') &&
-      !ref.current!.contains(event.target as HTMLElement)
-    ) {
+    if (PopupStack.isTopmost(ref.current!) && !ref.current!.contains(event.target as HTMLElement)) {
       onClose();
     }
   };
