+++ modules/popup/react/lib/usePopupStack.ts
@@ -1,22 +1,19 @@
 import React from 'react';
 
-import {PopupStack} from '@workday/canvas-kit-labs-react-popup-stack';
+import {PopupStack} from '@workday/canvas-kit-popup-stack';
 
 /**
  * Connects an element ref to the Popup stack
  * @param ref Ref of the element that is considered the container for a Popup. Usually the `Popup` component
  * @param type The type of popup
  */
-export const usePopupStack = <E extends HTMLElement>(
-  ref: React.RefObject<E>,
-  type: 'ephemeral' | 'persistent' = 'ephemeral'
-) => {
+export const usePopupStack = <E extends HTMLElement>(ref: React.RefObject<E>) => {
   // We useLayoutEffect to ensure proper timing of registration
   // of the element to the popup stack. Without this, the timing is unpredictable
   // when mixed with other frameworks. Other frameworks should also register as soon
   // as the element is available
   React.useLayoutEffect(() => {
-    PopupStack.add({element: ref.current!, type});
+    PopupStack.add({element: ref.current!});
 
     return () => {
       PopupStack.remove(ref.current!);
@@ -30,7 +27,7 @@ export const usePopupStack = <E extends HTMLElement>(
     isTopmost() {
       // If the type of popup is ephemeral, we only care about other ephemeral popups
       // If the type of popup is persistent, I'm not sure we care what's on top
-      return PopupStack.isTopmost(ref.current!, type === 'ephemeral' ? 'ephemeral' : undefined);
+      return PopupStack.isTopmost(ref.current!);
     },
   };
 };
