+++ modules/popup/react/package.json
@@ -48,15 +48,19 @@
     "@emotion/core": "^10.0.28",
     "@emotion/is-prop-valid": "^0.8.2",
     "@emotion/styled": "^10.0.27",
+    "@popperjs/core": "^2.1.1",
+    "@workday/canvas-kit-popup-stack": "0.0.0",
     "@workday/canvas-kit-react-button": "^4.0.0-beta.3",
     "@workday/canvas-kit-react-card": "^4.0.0-beta.3",
     "@workday/canvas-kit-react-common": "^4.0.0-beta.3",
     "@workday/canvas-kit-react-core": "^4.0.0-beta.3",
     "@workday/canvas-system-icons-web": "^1.0.20",
+    "focus-trap-js": "1.0.8",
     "uuid": "^3.3.3"
   },
   "devDependencies": {
-    "@workday/canvas-kit-labs-react-core": "^4.0.0-beta.3"
+    "@workday/canvas-kit-labs-react-core": "^4.0.0-beta.3",
+    "lodash": "^4.17.14"
   },
   "peerDependencies": {
     "react": ">= 16.8 < 17"
