+++ modules/tooltip/react/package.json
@@ -52,6 +52,7 @@
     "@emotion/styled": "^10.0.27",
     "@workday/canvas-kit-react-common": "^4.0.0-beta.3",
     "@workday/canvas-kit-react-core": "^4.0.0-beta.3",
+    "@workday/canvas-kit-react-popup": "^4.0.0-beta.3",
     "react-innertext": "^1.1.5",
     "uuid": "^3.3.3"
   },
